Revert "manage staff"
