+++ backend/index.js
@@ -5,7 +5,6 @@ import dotenv from "dotenv";
 import express from "express";
 import connectToDatabase from "./db/db.js";
 import authRouter from "./router/auth.js";
-import staffRouter from "./router/staff.js";
 
 dotenv.config(); // Load biến môi trường từ .env
 
@@ -15,7 +14,7 @@ const PORT = process.env.PORT || 5000;
 // Cấu hình CORS
 const corsOptions = {
   origin: "http://localhost:5173",
-  methods: ["GET", "POST", "PUT", "DELETE", "PATCH"],
+  methods: ["GET", "POST", "PUT", "DELETE"],
   credentials: true,
   allowedHeaders: ["Content-Type", "Authorization"],
 };
@@ -39,7 +38,6 @@ app.get("/", (req, res) => res.send("API working"));
 
 // Routes chính
 app.use("/api/auth", authRouter);
-app.use("/api/staff", staffRouter);
 
 // Kết nối DB và khởi chạy server
 const startServer = async () => {
